requirements.txt
ollama==0.6.1
python-dotenv==1.2.1

